Update cron job docs for time zone considerations
In discussion of https://github.com/harrystech/cronut/issues/11, it came to
light that differences between the server and user's reference time zone could
cause undesired behavior when cron jobs were pinged (for example, my server is
in UTC, but our time zone for reference is `America/New_York`). Adding a note about
specifying a time zone to the Cron Jobs section of the README.

Close #11.